third_party/WebKit/LayoutTests/http/tests/devtools/editor/text-editor-search-replace-expected.txt

Tests the search replace functionality.


Running: testReplaceAll
--- Before replace ---
// var a1, a2, a3;
const  a1, a2, a3;

--- After replace ---
//= var a$$, b$&, a3 /* $0 $1 $2 $& $$ \0 \1 */;
/** @const */ var a$$, b$&, a3 /* $0 $1 $2 $& $$ \0 \1 */;
